Attach to console, if it exists.
test-con3 is a Windows app, so normally has no 'console', and debug output
to 'stdout' goes nowhere.

But if the app is started in a 'console', a command prompt, attach to that
console and freopen conout$ so the ouput appears there.

And if an ouput redirection is added, then allow that to exist and avoid
the freopen which will break that redirection. ie bRedON=TRUE

	modified:   src/Win32/test-con3.cpp